--- tests/visual/character-creation.spec.ts
@@ -1,521 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-/**
- * Character Creation Flow Visual Regression Tests
- * 
- * These tests ensure visual consistency of the character creation wizard interface.
- * Focus on USER-FACING visual behavior for the multi-step character creation process.
- * 
- * Will FAIL initially (red phase) until:
- * - Playwright configuration exists
- * - Baseline screenshots are generated
- * - Character creation pages are accessible
- * 
- * Acceptance Criteria Validation:
- * - Character creation wizard layout remains visually consistent
- * - Attribute/skill assignment interfaces work correctly
- * - Portrait generation UI renders consistently
- * - Form validation states are visually stable
- */
-
-test.describe('Character Creation Wizard Visual Regression', () => {
-  test.beforeEach(async ({ page }) => {
-    // Navigate to character creation page
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Disable animations for consistent screenshots
-    await page.addStyleTag({
-      content: `
-        *, *::before, *::after {
-          animation-duration: 0s !important;
-          animation-delay: 0s !important;
-          transition-duration: 0s !important;
-          transition-delay: 0s !important;
-        }
-      `
-    });
-  });
-
-  test('should maintain character creation wizard layout consistency', async ({ page }) => {
-    // Core visual regression test for character creation interface
-    // Tests the overall wizard layout that users interact with
-    
-    await expect(page).toHaveScreenshot('character-creation-wizard-layout.png', {
-      fullPage: true,
-      animations: 'disabled',
-      threshold: 0.3
-    });
-  });
-
-  test('should maintain character basic info step visual layout', async ({ page }) => {
-    // Test the basic information step (name, description, etc.)
-    // This is typically the first step users encounter
-    
-    const basicInfoForm = page.locator('form, .character-form, .basic-info-step, main').first();
-    await expect(basicInfoForm).toHaveScreenshot('character-creation-basic-info.png', {
-      threshold: 0.25
-    });
-  });
-
-  test('should maintain character name input visual consistency', async ({ page }) => {
-    // Test character name input field visual states
-    // Critical for character identity creation
-    
-    const nameInput = page.locator('input[name*="name"], input[placeholder*="name"], #name, #characterName').first();
-    
-    if (await nameInput.count() > 0) {
-      // Default state
-      await expect(nameInput).toHaveScreenshot('character-creation-name-input-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await nameInput.focus();
-      await expect(nameInput).toHaveScreenshot('character-creation-name-input-focus.png', {
-        threshold: 0.3
-      });
-      
-      // With content
-      await nameInput.fill('Test Character');
-      await expect(nameInput).toHaveScreenshot('character-creation-name-input-filled.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain character description area visual consistency', async ({ page }) => {
-    // Test character description/background textarea
-    const descriptionArea = page.locator('textarea, [name*="description"], [name*="background"]').first();
-    
-    if (await descriptionArea.count() > 0) {
-      // Default state
-      await expect(descriptionArea).toHaveScreenshot('character-creation-description-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await descriptionArea.focus();
-      await expect(descriptionArea).toHaveScreenshot('character-creation-description-focus.png', {
-        threshold: 0.3
-      });
-      
-      // With content
-      await descriptionArea.fill('A brave adventurer seeking glory.');
-      await expect(descriptionArea).toHaveScreenshot('character-creation-description-filled.png', {
-        threshold: 0.3
-      });
-    }
-  });
-});
-
-test.describe('Character Attributes/Skills Visual Regression', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Try to navigate to attributes step if wizard exists
-    const nextButton = page.locator('button:has-text("Next"), [data-testid="next-button"]').first();
-    if (await nextButton.count() > 0) {
-      // Fill required fields first
-      const nameInput = page.locator('input[name*="name"], input[placeholder*="name"]').first();
-      if (await nameInput.count() > 0) {
-        await nameInput.fill('Test Character');
-        await nextButton.click();
-        await page.waitForTimeout(200);
-      }
-    }
-    
-    // Disable animations
-    await page.addStyleTag({
-      content: `
-        *, *::before, *::after {
-          animation-duration: 0s !important;
-          transition-duration: 0s !important;
-        }
-      `
-    });
-  });
-
-  test('should maintain attribute assignment interface visual consistency', async ({ page }) => {
-    // Test attribute point assignment interface
-    // Critical for character stat allocation
-    
-    const attributeSection = page.locator(
-      '[data-testid="attributes"], .attributes, .stats, .character-attributes'
-    ).first();
-    
-    if (await attributeSection.count() > 0) {
-      await expect(attributeSection).toHaveScreenshot('character-creation-attributes-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for any slider, input, or number controls
-      const attributeControls = page.locator('input[type="range"], input[type="number"], .slider').first();
-      if (await attributeControls.count() > 0) {
-        await expect(attributeControls).toHaveScreenshot('character-creation-attribute-controls.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain point pool visual consistency', async ({ page }) => {
-    // Test point allocation pool display
-    // Important for showing available/remaining points
-    
-    const pointPool = page.locator(
-      '[data-testid="point-pool"], .point-pool, .points-remaining, .available-points'
-    ).first();
-    
-    if (await pointPool.count() > 0) {
-      await expect(pointPool).toHaveScreenshot('character-creation-point-pool.png', {
-        threshold: 0.25
-      });
-    } else {
-      // Look for any text showing points or pool information
-      const pointText = page.locator('text=/points|remaining|available/i').first();
-      if (await pointText.count() > 0) {
-        await expect(pointText.locator('..').first()).toHaveScreenshot('character-creation-points-display.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-
-  test('should maintain attribute slider visual states', async ({ page }) => {
-    // Test attribute slider interactions
-    const slider = page.locator('input[type="range"], .slider').first();
-    
-    if (await slider.count() > 0) {
-      // Default state
-      await expect(slider).toHaveScreenshot('character-creation-slider-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await slider.focus();
-      await expect(slider).toHaveScreenshot('character-creation-slider-focus.png', {
-        threshold: 0.3
-      });
-      
-      // Modified value
-      await slider.fill('75');
-      await expect(slider).toHaveScreenshot('character-creation-slider-modified.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain skill selection visual consistency', async ({ page }) => {
-    // Test skill selection interface
-    const skillSection = page.locator(
-      '[data-testid="skills"], .skills, .character-skills, .skill-selection'
-    ).first();
-    
-    if (await skillSection.count() > 0) {
-      await expect(skillSection).toHaveScreenshot('character-creation-skills-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for checkboxes or selection lists that might be skills
-      const skillControls = page.locator('input[type="checkbox"], .skill-item, .selection-list').first();
-      if (await skillControls.count() > 0) {
-        await expect(skillControls).toHaveScreenshot('character-creation-skill-controls.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-});
-
-test.describe('Character Portrait Visual Regression', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Navigate to portrait step if possible
-    const nextButton = page.locator('button:has-text("Next")').first();
-    if (await nextButton.count() > 0) {
-      // Fill basic info and navigate
-      const nameInput = page.locator('input[name*="name"]').first();
-      if (await nameInput.count() > 0) {
-        await nameInput.fill('Test Character');
-        await nextButton.click();
-        await page.waitForTimeout(200);
-        
-        // Try to go to next step (might be portrait)
-        if (await nextButton.count() > 0) {
-          await nextButton.click();
-          await page.waitForTimeout(200);
-        }
-      }
-    }
-  });
-
-  test('should maintain portrait generation interface visual consistency', async ({ page }) => {
-    // Test AI portrait generation interface
-    // Important for character visual representation
-    
-    const portraitSection = page.locator(
-      '[data-testid="portrait"], .portrait, .character-image, .image-generation'
-    ).first();
-    
-    if (await portraitSection.count() > 0) {
-      await expect(portraitSection).toHaveScreenshot('character-creation-portrait-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for any image-related elements
-      const imageElements = page.locator('img, .image-placeholder, [role="img"]');
-      if (await imageElements.count() > 0) {
-        await expect(imageElements.first()).toHaveScreenshot('character-creation-image-element.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain portrait placeholder visual consistency', async ({ page }) => {
-    // Test portrait placeholder before generation
-    const placeholder = page.locator(
-      '.portrait-placeholder, .image-placeholder, .no-image, [data-testid="portrait-placeholder"]'
-    ).first();
-    
-    if (await placeholder.count() > 0) {
-      await expect(placeholder).toHaveScreenshot('character-creation-portrait-placeholder.png', {
-        threshold: 0.25
-      });
-    }
-  });
-
-  test('should maintain portrait generation controls visual consistency', async ({ page }) => {
-    // Test controls for generating/uploading portraits
-    const generateButton = page.locator(
-      'button:has-text("Generate"), [data-testid="generate-portrait"], .generate-image'
-    ).first();
-    
-    const uploadButton = page.locator(
-      'button:has-text("Upload"), [data-testid="upload-portrait"], input[type="file"]'
-    ).first();
-    
-    if (await generateButton.count() > 0) {
-      await expect(generateButton).toHaveScreenshot('character-creation-generate-button.png', {
-        threshold: 0.2
-      });
-    }
-    
-    if (await uploadButton.count() > 0) {
-      await expect(uploadButton).toHaveScreenshot('character-creation-upload-button.png', {
-        threshold: 0.2
-      });
-    }
-  });
-
-  test('should maintain portrait loading state visual consistency', async ({ page }) => {
-    // Test loading state during portrait generation
-    const generateButton = page.locator('button:has-text("Generate"), [data-testid="generate-portrait"]').first();
-    
-    if (await generateButton.count() > 0) {
-      // Mock the portrait generation request
-      await page.route('**/api/generate-portrait', route => {
-        setTimeout(() => {
-          route.fulfill({
-            status: 200,
-            contentType: 'application/json',
-            body: JSON.stringify({ portraitUrl: 'data:image/svg+xml;base64,PHN2Zz48L3N2Zz4=' })
-          });
-        }, 1000);
-      });
-      
-      await generateButton.click();
-      
-      // Capture loading state
-      const loadingElement = page.locator('.loading, .spinner, [data-testid="loading"]').first();
-      if (await loadingElement.isVisible()) {
-        await expect(loadingElement).toHaveScreenshot('character-creation-portrait-loading.png', {
-          threshold: 0.4
-        });
-      }
-    }
-  });
-});
-
-test.describe('Character Creation Validation Visual Tests', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-  });
-
-  test('should maintain validation error visual consistency', async ({ page }) => {
-    // Test visual appearance of validation errors
-    // Critical for user feedback during character creation
-    
-    // Try to submit without required data
-    const submitButton = page.locator(
-      'button[type="submit"], button:has-text("Create"), button:has-text("Finish")'
-    ).first();
-    
-    if (await submitButton.count() > 0) {
-      await submitButton.click();
-      await page.waitForTimeout(100);
-      
-      // Look for validation messages
-      const errorMessages = page.locator('.error, [role="alert"], .validation-error, .field-error');
-      
-      if (await errorMessages.count() > 0) {
-        await expect(errorMessages.first()).toHaveScreenshot('character-creation-validation-error.png', {
-          threshold: 0.25
-        });
-      }
-      
-      // Capture form with errors
-      await expect(page).toHaveScreenshot('character-creation-form-with-errors.png', {
-        fullPage: true,
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain point allocation validation visual feedback', async ({ page }) => {
-    // Test validation for point over-allocation
-    const slider = page.locator('input[type="range"], input[type="number"]').first();
-    
-    if (await slider.count() > 0) {
-      // Try to set an invalid value
-      await slider.fill('999');
-      await page.keyboard.press('Tab'); // Trigger validation
-      
-      // Look for validation feedback
-      const validation = page.locator('.error, .invalid, [aria-invalid="true"]');
-      if (await validation.count() > 0) {
-        await expect(validation.first()).toHaveScreenshot('character-creation-point-validation.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain required field indicators visual consistency', async ({ page }) => {
-    // Test visual indicators for required fields
-    const requiredFields = page.locator('input[required], textarea[required], select[required]');
-    const requiredIndicators = page.locator('.required, [aria-required="true"]');
-    
-    if (await requiredFields.count() > 0 || await requiredIndicators.count() > 0) {
-      const firstRequired = await requiredFields.count() > 0 
-        ? requiredFields.first() 
-        : requiredIndicators.first();
-        
-      await expect(firstRequired).toHaveScreenshot('character-creation-required-field.png', {
-        threshold: 0.2
-      });
-    }
-  });
-});
-
-test.describe('Character Creation Multi-Step Flow Visual Tests', () => {
-  test('should maintain visual consistency across character creation steps', async ({ page }) => {
-    // Test the complete multi-step character creation flow
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Capture initial step
-    await expect(page).toHaveScreenshot('character-creation-step-1.png', {
-      fullPage: true,
-      threshold: 0.3
-    });
-    
-    // Try to progress through steps
-    const nextButton = page.locator('button:has-text("Next"), [data-testid="next-button"]').first();
-    
-    if (await nextButton.count() > 0) {
-      // Fill minimal required data
-      const nameInput = page.locator('input[name*="name"]').first();
-      if (await nameInput.count() > 0) {
-        await nameInput.fill('Test Character');
-      }
-      
-      await nextButton.click();
-      await page.waitForTimeout(200);
-      
-      // Capture second step
-      await expect(page).toHaveScreenshot('character-creation-step-2.png', {
-        fullPage: true,
-        threshold: 0.3
-      });
-      
-      // Try to go to third step
-      if (await nextButton.count() > 0) {
-        await nextButton.click();
-        await page.waitForTimeout(200);
-        
-        await expect(page).toHaveScreenshot('character-creation-step-3.png', {
-          fullPage: true,
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain character creation progress indicator', async ({ page }) => {
-    // Test progress indicator throughout character creation
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-    
-    const progressElement = page.locator(
-      '[role="progressbar"], .progress, .step-indicator, .wizard-steps'
-    ).first();
-    
-    if (await progressElement.count() > 0) {
-      await expect(progressElement).toHaveScreenshot('character-creation-progress-step-1.png', {
-        threshold: 0.25
-      });
-      
-      // Advance and check progress update
-      const nextButton = page.locator('button:has-text("Next")').first();
-      if (await nextButton.count() > 0) {
-        const nameInput = page.locator('input').first();
-        if (await nameInput.count() > 0) {
-          await nameInput.fill('Test');
-        }
-        
-        await nextButton.click();
-        await page.waitForTimeout(200);
-        
-        await expect(progressElement).toHaveScreenshot('character-creation-progress-step-2.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-
-  test('should maintain character creation summary/review visual layout', async ({ page }) => {
-    // Test final review/summary step if it exists
-    await page.goto('/characters/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Try to navigate to final step by filling minimal data and progressing
-    const nextButton = page.locator('button:has-text("Next")');
-    const nameInput = page.locator('input[name*="name"]').first();
-    
-    if (await nameInput.count() > 0 && await nextButton.count() > 0) {
-      await nameInput.fill('Test Character');
-      
-      // Try to reach final step
-      for (let i = 0; i < 3; i++) {
-        if (await nextButton.count() > 0) {
-          await nextButton.click();
-          await page.waitForTimeout(200);
-        }
-      }
-      
-      // Look for summary or review content
-      const summaryContent = page.locator('.summary, .review, .character-preview, .final-step');
-      if (await summaryContent.count() > 0) {
-        await expect(summaryContent.first()).toHaveScreenshot('character-creation-summary.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-});
