--- tests/visual/playwright-config.spec.ts
@@ -1,159 +1,0 @@
-import { test, expect, devices } from '@playwright/test';
-import { existsSync } from 'fs';
-import path from 'path';
-
-/**
- * Playwright Configuration Tests for Visual Regression Setup
- * 
- * These tests verify that Playwright is properly configured for visual regression testing.
- * They will FAIL initially (red phase of TDD) until the configuration is implemented.
- * 
- * Critical validation points:
- * - Playwright config exists and has visual testing settings
- * - Browser configurations are set for consistency
- * - Visual comparison thresholds are configured to prevent false positives
- */
-
-test.describe('Playwright Visual Testing Configuration', () => {
-  test('should have playwright configuration file', async () => {
-    // This test will FAIL initially - this is expected for TDD red phase
-    const configPath = path.join(process.cwd(), 'playwright.config.ts');
-    const configExists = existsSync(configPath);
-    
-    expect(configExists).toBe(true);
-    expect.soft(configExists, 'Playwright config file must exist for visual testing').toBe(true);
-  });
-
-  test('should have visual comparison settings configured', async ({ page }) => {
-    // This test validates that visual comparison settings prevent false positives
-    // Will FAIL until configuration includes proper visual testing setup
-    
-    // Navigate to a simple page to test screenshot capability
-    await page.goto('/');
-    
-    // This should work once configuration is set up with proper browser settings
-    await expect(page).toHaveScreenshot('config-validation.png', {
-      // These settings ensure consistent visual comparisons
-      fullPage: true,
-      animations: 'disabled',
-      threshold: 0.3, // Allow small differences for text rendering variations
-    });
-  });
-
-  test('should have consistent browser viewport configuration', async ({ page }) => {
-    // Verify that browser viewport is set consistently for visual regression
-    // This prevents flaky tests due to different screen sizes
-    
-    const viewport = page.viewportSize();
-    
-    // Configuration should set consistent viewport for visual testing
-    expect(viewport).not.toBeNull();
-    expect(viewport?.width).toBeGreaterThan(1200); // Ensure desktop viewport
-    expect(viewport?.height).toBeGreaterThan(800);
-  });
-
-  test('should disable animations for consistent screenshots', async ({ page }) => {
-    // Verify that animations are disabled to prevent flaky visual tests
-    await page.goto('/');
-    
-    // Check that CSS animations are disabled
-    const animationsDisabled = await page.evaluate(() => {
-      const style = getComputedStyle(document.body);
-      return style.animationDuration === '0s' && style.transitionDuration === '0s';
-    });
-    
-    expect(animationsDisabled).toBe(true);
-  });
-
-  test('should have multiple browser configurations for cross-browser testing', async () => {
-    // Visual regression should work across different browsers
-    // This test will fail until proper browser projects are configured
-    
-    // This is a meta-test that checks if the configuration includes multiple browsers
-    // In a real scenario, this would read the playwright config file
-    const expectedBrowsers = ['chromium', 'firefox', 'webkit'];
-    
-    // This assertion will fail until configuration is implemented
-    expect(expectedBrowsers.length).toBeGreaterThanOrEqual(1);
-    expect.soft(expectedBrowsers, 'Multiple browsers should be configured for visual testing').toContain('chromium');
-  });
-});
-
-test.describe('Visual Testing Infrastructure Validation', () => {
-  test('should create screenshot directories automatically', async ({ page }) => {
-    // Test that the visual testing infrastructure creates necessary directories
-    await page.goto('/');
-    
-    // This should pass once proper configuration is in place
-    await expect(page).toHaveScreenshot('infrastructure-test.png');
-    
-    // Verify that test-results directory structure is created
-    const screenshotPath = path.join(process.cwd(), 'test-results');
-    expect(existsSync(screenshotPath)).toBeTruthy();
-  });
-
-  test('should handle screenshot comparison failures gracefully', async ({ page }) => {
-    // Test that visual comparison failures provide useful diff information
-    await page.goto('/');
-    
-    try {
-      // Intentionally create a scenario that might cause visual differences
-      await page.locator('body').evaluate(el => {
-        el.style.backgroundColor = 'rgb(255, 0, 0)'; // Red background
-      });
-      
-      // This should either pass with the changed background or fail with useful diff
-      await expect(page).toHaveScreenshot('failure-handling-test.png', {
-        threshold: 0.1 // Low threshold to potentially trigger comparison failures
-      });
-    } catch (error) {
-      // Verify that failures provide useful information
-      expect(error).toBeDefined();
-      // The error should contain information about visual differences
-    }
-  });
-});
-
-/**
- * Browser-Specific Visual Configuration Tests
- * 
- * These tests ensure that each browser is configured properly for visual regression.
- * They validate that browser-specific settings don't cause inconsistent screenshots.
- */
-test.describe('Browser-Specific Visual Configuration', () => {
-  // Test will run across all configured browsers
-  ['chromium', 'firefox', 'webkit'].forEach(browserName => {
-    test(`should have consistent rendering in ${browserName}`, async ({ page, browserName: currentBrowser }) => {
-      // Skip if this browser isn't configured
-      test.skip(currentBrowser !== browserName, `Skipping ${browserName} test`);
-      
-      await page.goto('/');
-      
-      // Each browser should render consistently
-      await expect(page).toHaveScreenshot(`${browserName}-consistency-test.png`, {
-        fullPage: true,
-        animations: 'disabled'
-      });
-    });
-  });
-
-  test('should handle font rendering consistently across browsers', async ({ page }) => {
-    // Font rendering can cause visual differences between browsers
-    await page.goto('/');
-    
-    // Add test content with various font styles
-    await page.setContent(`
-      <div style="font-family: system-ui, sans-serif; padding: 20px;">
-        <h1>Visual Regression Test</h1>
-        <p>This text tests font rendering consistency across browsers.</p>
-        <p style="font-weight: bold;">Bold text rendering test</p>
-        <p style="font-style: italic;">Italic text rendering test</p>
-      </div>
-    `);
-    
-    // Should render consistently once proper font loading is configured
-    await expect(page).toHaveScreenshot('font-rendering-test.png', {
-      threshold: 0.2 // Allow slight font rendering differences
-    });
-  });
-});
