--- tests/visual/settings.spec.ts
@@ -1,527 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-/**
- * Settings Page Layout Visual Regression Tests
- * 
- * These tests ensure visual consistency of the settings/preferences interface.
- * Focus on USER-FACING visual behavior for configuration and preferences.
- * 
- * Will FAIL initially (red phase) until:
- * - Playwright configuration exists
- * - Baseline screenshots are generated
- * - Settings pages are accessible
- * 
- * Acceptance Criteria Validation:
- * - Settings page layout remains visually consistent
- * - Form controls and inputs render correctly
- * - Sections and organization are visually stable
- * - Save/reset functionality UI is consistent
- */
-
-test.describe('Settings Page Visual Regression', () => {
-  test.beforeEach(async ({ page }) => {
-    // Navigate to settings page
-    await page.goto('/settings');
-    await page.waitForLoadState('networkidle');
-    
-    // Disable animations for consistent screenshots
-    await page.addStyleTag({
-      content: `
-        *, *::before, *::after {
-          animation-duration: 0s !important;
-          animation-delay: 0s !important;
-          transition-duration: 0s !important;
-          transition-delay: 0s !important;
-        }
-      `
-    });
-  });
-
-  test('should maintain settings page layout consistency', async ({ page }) => {
-    // Core visual regression test for settings interface
-    // Tests the overall settings layout that users configure
-    
-    await expect(page).toHaveScreenshot('settings-page-layout.png', {
-      fullPage: true,
-      animations: 'disabled',
-      threshold: 0.3
-    });
-  });
-
-  test('should maintain settings navigation/sections visual consistency', async ({ page }) => {
-    // Test settings navigation or section headers
-    // Important for organizing different setting categories
-    
-    const settingsNav = page.locator(
-      '[data-testid="settings-nav"], .settings-nav, .settings-tabs, .settings-sections'
-    ).first();
-    
-    if (await settingsNav.count() > 0) {
-      await expect(settingsNav).toHaveScreenshot('settings-navigation.png', {
-        threshold: 0.25
-      });
-    } else {
-      // Look for section headers or tabs
-      const sectionHeaders = page.locator('h2, h3, .section-header, [role="tab"]').first();
-      if (await sectionHeaders.count() > 0) {
-        await expect(sectionHeaders.locator('..').first()).toHaveScreenshot('settings-section-headers.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-
-  test('should maintain settings form controls visual consistency', async ({ page }) => {
-    // Test various form controls (inputs, selects, checkboxes, etc.)
-    // Critical for user preference configuration
-    
-    const formSection = page.locator('form, .settings-form, .preferences-form').first();
-    
-    if (await formSection.count() > 0) {
-      await expect(formSection).toHaveScreenshot('settings-form-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for any form controls
-      const controls = page.locator('input, select, textarea, [role="checkbox"], [role="slider"]').first();
-      if (await controls.count() > 0) {
-        await expect(controls.locator('..').first()).toHaveScreenshot('settings-form-controls.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain settings save/reset buttons visual consistency', async ({ page }) => {
-    // Test action buttons for saving or resetting preferences
-    const saveButton = page.locator('button:has-text("Save"), [data-testid="save-settings"]').first();
-    const resetButton = page.locator('button:has-text("Reset"), button:has-text("Default"), [data-testid="reset-settings"]').first();
-    
-    if (await saveButton.count() > 0) {
-      await expect(saveButton).toHaveScreenshot('settings-save-button.png', {
-        threshold: 0.2
-      });
-    }
-    
-    if (await resetButton.count() > 0) {
-      await expect(resetButton).toHaveScreenshot('settings-reset-button.png', {
-        threshold: 0.2
-      });
-    }
-  });
-});
-
-test.describe('Settings Form Input Visual Tests', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/settings');
-    await page.waitForLoadState('networkidle');
-    
-    // Disable animations
-    await page.addStyleTag({
-      content: `
-        *, *::before, *::after {
-          animation-duration: 0s !important;
-          transition-duration: 0s !important;
-        }
-      `
-    });
-  });
-
-  test('should maintain text input visual states', async ({ page }) => {
-    // Test text input controls in settings
-    const textInput = page.locator('input[type="text"], input[type="email"], input:not([type])').first();
-    
-    if (await textInput.count() > 0) {
-      // Default state
-      await expect(textInput).toHaveScreenshot('settings-text-input-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await textInput.focus();
-      await expect(textInput).toHaveScreenshot('settings-text-input-focus.png', {
-        threshold: 0.3
-      });
-      
-      // With content
-      await textInput.fill('Test Setting Value');
-      await expect(textInput).toHaveScreenshot('settings-text-input-filled.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain checkbox visual states', async ({ page }) => {
-    // Test checkbox controls for boolean settings
-    const checkbox = page.locator('input[type="checkbox"], [role="checkbox"]').first();
-    
-    if (await checkbox.count() > 0) {
-      // Unchecked state
-      await expect(checkbox).toHaveScreenshot('settings-checkbox-unchecked.png', {
-        threshold: 0.2
-      });
-      
-      // Checked state
-      await checkbox.check();
-      await expect(checkbox).toHaveScreenshot('settings-checkbox-checked.png', {
-        threshold: 0.3
-      });
-      
-      // Focus state
-      await checkbox.focus();
-      await expect(checkbox).toHaveScreenshot('settings-checkbox-focus.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain select dropdown visual states', async ({ page }) => {
-    // Test select dropdown controls
-    const select = page.locator('select, [role="combobox"]').first();
-    
-    if (await select.count() > 0) {
-      // Default state
-      await expect(select).toHaveScreenshot('settings-select-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await select.focus();
-      await expect(select).toHaveScreenshot('settings-select-focus.png', {
-        threshold: 0.3
-      });
-      
-      // Try to open dropdown (if supported)
-      if (await select.locator('option').count() > 1) {
-        await select.selectOption({ index: 1 });
-        await expect(select).toHaveScreenshot('settings-select-selected.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain range slider visual states', async ({ page }) => {
-    // Test range slider controls for numeric settings
-    const slider = page.locator('input[type="range"], [role="slider"]').first();
-    
-    if (await slider.count() > 0) {
-      // Default state
-      await expect(slider).toHaveScreenshot('settings-slider-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await slider.focus();
-      await expect(slider).toHaveScreenshot('settings-slider-focus.png', {
-        threshold: 0.3
-      });
-      
-      // Modified value
-      await slider.fill('75');
-      await expect(slider).toHaveScreenshot('settings-slider-modified.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain textarea visual states', async ({ page }) => {
-    // Test textarea controls for longer text settings
-    const textarea = page.locator('textarea').first();
-    
-    if (await textarea.count() > 0) {
-      // Default state
-      await expect(textarea).toHaveScreenshot('settings-textarea-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await textarea.focus();
-      await expect(textarea).toHaveScreenshot('settings-textarea-focus.png', {
-        threshold: 0.3
-      });
-      
-      // With content
-      await textarea.fill('This is a longer text setting that might wrap to multiple lines in the textarea.');
-      await expect(textarea).toHaveScreenshot('settings-textarea-filled.png', {
-        threshold: 0.3
-      });
-    }
-  });
-});
-
-test.describe('Settings Sections Visual Organization', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/settings');
-    await page.waitForLoadState('networkidle');
-  });
-
-  test('should maintain user profile settings visual layout', async ({ page }) => {
-    // Test user profile/account settings section
-    const profileSection = page.locator(
-      '[data-testid="profile-settings"], .profile-settings, .user-settings, .account-settings'
-    ).first();
-    
-    if (await profileSection.count() > 0) {
-      await expect(profileSection).toHaveScreenshot('settings-profile-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for profile-related fields
-      const profileFields = page.locator('input[name*="name"], input[name*="email"], input[name*="username"]').first();
-      if (await profileFields.count() > 0) {
-        await expect(profileFields.locator('..').first()).toHaveScreenshot('settings-profile-fields.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain game preferences visual layout', async ({ page }) => {
-    // Test game-specific preference settings
-    const gameSettings = page.locator(
-      '[data-testid="game-settings"], .game-settings, .game-preferences, .gameplay-settings'
-    ).first();
-    
-    if (await gameSettings.count() > 0) {
-      await expect(gameSettings).toHaveScreenshot('settings-game-preferences.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for game-related settings
-      const gameFields = page.locator('input[name*="game"], input[name*="difficulty"], select[name*="theme"]').first();
-      if (await gameFields.count() > 0) {
-        await expect(gameFields.locator('..').first()).toHaveScreenshot('settings-game-fields.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain accessibility settings visual layout', async ({ page }) => {
-    // Test accessibility/UI preference settings
-    const accessibilitySettings = page.locator(
-      '[data-testid="accessibility-settings"], .accessibility-settings, .ui-settings, .display-settings'
-    ).first();
-    
-    if (await accessibilitySettings.count() > 0) {
-      await expect(accessibilitySettings).toHaveScreenshot('settings-accessibility-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for accessibility-related controls
-      const a11yFields = page.locator('input[name*="theme"], input[name*="font"], input[name*="contrast"]').first();
-      if (await a11yFields.count() > 0) {
-        await expect(a11yFields.locator('..').first()).toHaveScreenshot('settings-accessibility-fields.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain notification settings visual layout', async ({ page }) => {
-    // Test notification preference settings
-    const notificationSettings = page.locator(
-      '[data-testid="notification-settings"], .notification-settings, .alerts-settings'
-    ).first();
-    
-    if (await notificationSettings.count() > 0) {
-      await expect(notificationSettings).toHaveScreenshot('settings-notification-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for notification-related checkboxes
-      const notificationFields = page.locator('input[name*="notification"], input[name*="alert"], input[name*="email"]');
-      if (await notificationFields.count() > 0) {
-        await expect(notificationFields.first().locator('..').first()).toHaveScreenshot('settings-notification-fields.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-});
-
-test.describe('Settings Validation and Feedback Visual Tests', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/settings');
-    await page.waitForLoadState('networkidle');
-  });
-
-  test('should maintain validation error visual consistency', async ({ page }) => {
-    // Test visual appearance of validation errors in settings
-    const textInput = page.locator('input[type="text"], input[type="email"]').first();
-    
-    if (await textInput.count() > 0) {
-      // Enter invalid data to trigger validation
-      await textInput.fill('invalid-email');
-      await textInput.blur(); // Trigger validation
-      
-      // Look for validation errors
-      const errorMessage = page.locator('.error, [role="alert"], .validation-error, .field-error').first();
-      if (await errorMessage.count() > 0) {
-        await expect(errorMessage).toHaveScreenshot('settings-validation-error.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-
-  test('should maintain save success feedback visual consistency', async ({ page }) => {
-    // Test success feedback after saving settings
-    const saveButton = page.locator('button:has-text("Save"), [data-testid="save-settings"]').first();
-    
-    if (await saveButton.count() > 0) {
-      // Mock successful save response
-      await page.route('**/api/settings', route => {
-        route.fulfill({
-          status: 200,
-          contentType: 'application/json',
-          body: JSON.stringify({ success: true })
-        });
-      });
-      
-      await saveButton.click();
-      await page.waitForTimeout(200);
-      
-      // Look for success feedback
-      const successMessage = page.locator('.success, .saved, [data-testid="success-message"]').first();
-      if (await successMessage.count() > 0) {
-        await expect(successMessage).toHaveScreenshot('settings-save-success.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-
-  test('should maintain unsaved changes indicator visual consistency', async ({ page }) => {
-    // Test indicator for unsaved changes
-    const textInput = page.locator('input[type="text"]').first();
-    
-    if (await textInput.count() > 0) {
-      await textInput.fill('Modified value');
-      await page.waitForTimeout(100);
-      
-      // Look for unsaved changes indicator
-      const unsavedIndicator = page.locator('.unsaved, .modified, [data-testid="unsaved-changes"]').first();
-      if (await unsavedIndicator.count() > 0) {
-        await expect(unsavedIndicator).toHaveScreenshot('settings-unsaved-indicator.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-});
-
-test.describe('Settings Responsive Visual Tests', () => {
-  const viewports = [
-    { name: 'mobile', width: 375, height: 667 },
-    { name: 'tablet', width: 768, height: 1024 },
-    { name: 'desktop', width: 1200, height: 800 }
-  ];
-
-  viewports.forEach(({ name, width, height }) => {
-    test(`should maintain settings visual consistency on ${name} viewport`, async ({ page }) => {
-      await page.setViewportSize({ width, height });
-      await page.goto('/settings');
-      await page.waitForLoadState('networkidle');
-      
-      // Disable animations
-      await page.addStyleTag({
-        content: `
-          *, *::before, *::after {
-            animation-duration: 0s !important;
-            transition-duration: 0s !important;
-          }
-        `
-      });
-      
-      await expect(page).toHaveScreenshot(`settings-${name}-viewport.png`, {
-        fullPage: true,
-        animations: 'disabled',
-        threshold: 0.3
-      });
-    });
-  });
-
-  test('should maintain mobile settings navigation layout', async ({ page }) => {
-    // Test mobile-specific settings navigation
-    await page.setViewportSize({ width: 375, height: 667 });
-    await page.goto('/settings');
-    await page.waitForLoadState('networkidle');
-    
-    const mobileNav = page.locator('.mobile-nav, .settings-mobile-menu, .collapsed-nav').first();
-    if (await mobileNav.count() > 0) {
-      await expect(mobileNav).toHaveScreenshot('settings-mobile-navigation.png', {
-        threshold: 0.3
-      });
-    }
-  });
-});
-
-test.describe('Settings Empty and Loading States Visual Tests', () => {
-  test('should maintain settings loading state visual consistency', async ({ page }) => {
-    // Test loading state while fetching settings
-    await page.route('**/api/settings', route => {
-      setTimeout(() => {
-        route.fulfill({
-          status: 200,
-          contentType: 'application/json',
-          body: JSON.stringify({ theme: 'dark', notifications: true })
-        });
-      }, 1000);
-    });
-    
-    await page.goto('/settings');
-    
-    // Capture loading state
-    const loadingElement = page.locator('.loading, .spinner, [data-testid="loading"]').first();
-    if (await loadingElement.isVisible()) {
-      await expect(loadingElement).toHaveScreenshot('settings-loading-state.png', {
-        threshold: 0.4
-      });
-    }
-    
-    // Wait for content to load and capture final state
-    await page.waitForLoadState('networkidle');
-    await expect(page).toHaveScreenshot('settings-loaded-state.png', {
-      fullPage: true,
-      threshold: 0.3
-    });
-  });
-
-  test('should maintain settings error state visual consistency', async ({ page }) => {
-    // Test error state when settings fail to load
-    await page.route('**/api/settings', route => {
-      route.abort('failed');
-    });
-    
-    await page.goto('/settings');
-    await page.waitForTimeout(500);
-    
-    // Look for error messages
-    const errorElement = page.locator('.error, [role="alert"], .error-message').first();
-    if (await errorElement.count() > 0) {
-      await expect(errorElement).toHaveScreenshot('settings-error-message.png', {
-        threshold: 0.25
-      });
-    }
-    
-    // Capture page with error state
-    await expect(page).toHaveScreenshot('settings-with-error.png', {
-      fullPage: true,
-      threshold: 0.3
-    });
-  });
-
-  test('should maintain default settings visual layout', async ({ page }) => {
-    // Test visual appearance with default/empty settings
-    await page.goto('/settings');
-    await page.waitForLoadState('networkidle');
-    
-    // Capture the default state of settings
-    await expect(page).toHaveScreenshot('settings-default-state.png', {
-      fullPage: true,
-      threshold: 0.3
-    });
-  });
-});
