--- tests/visual/world-creation.spec.ts
@@ -1,383 +1,0 @@
-import { test, expect } from '@playwright/test';
-
-/**
- * World Creation Interface Visual Regression Tests
- * 
- * These tests ensure visual consistency of the world creation wizard interface.
- * Focus on USER-FACING visual behavior for the multi-step world creation process.
- * 
- * Will FAIL initially (red phase) until:
- * - Playwright configuration exists
- * - Baseline screenshots are generated
- * - World creation pages are accessible
- * 
- * Acceptance Criteria Validation:
- * - World creation wizard layout remains visually consistent
- * - Step progression indicators work correctly
- * - Form elements and validation messages are visually stable
- * - Image generation UI renders consistently
- */
-
-test.describe('World Creation Wizard Visual Regression', () => {
-  test.beforeEach(async ({ page }) => {
-    // Navigate to world creation page
-    await page.goto('/world/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Disable animations for consistent screenshots
-    await page.addStyleTag({
-      content: `
-        *, *::before, *::after {
-          animation-duration: 0s !important;
-          animation-delay: 0s !important;
-          transition-duration: 0s !important;
-          transition-delay: 0s !important;
-        }
-      `
-    });
-  });
-
-  test('should maintain world creation wizard layout consistency', async ({ page }) => {
-    // Core visual regression test for world creation interface
-    // Tests the overall wizard layout that users interact with
-    
-    await expect(page).toHaveScreenshot('world-creation-wizard-layout.png', {
-      fullPage: true,
-      animations: 'disabled',
-      threshold: 0.3
-    });
-  });
-
-  test('should maintain step indicator visual consistency', async ({ page }) => {
-    // Test step progression indicators (breadcrumbs, progress bar, etc.)
-    // Critical for user navigation through the wizard
-    
-    const stepIndicator = page.locator(
-      '[data-testid="step-indicator"], .step-indicator, .progress, .breadcrumb, [role="progressbar"]'
-    ).first();
-    
-    if (await stepIndicator.count() > 0) {
-      await expect(stepIndicator).toHaveScreenshot('world-creation-step-indicator.png', {
-        threshold: 0.2
-      });
-    } else {
-      // Look for any navigation or progress elements in the header/top area
-      const headerArea = page.locator('header, nav, .wizard-header').first();
-      if (await headerArea.count() > 0) {
-        await expect(headerArea).toHaveScreenshot('world-creation-navigation.png', {
-          threshold: 0.2
-        });
-      }
-    }
-  });
-
-  test('should maintain basic info step visual layout', async ({ page }) => {
-    // Test the first step of world creation (typically basic information)
-    // This is usually the entry point users see
-    
-    const mainForm = page.locator('form, main, .wizard-step').first();
-    await expect(mainForm).toHaveScreenshot('world-creation-basic-info-step.png', {
-      threshold: 0.25
-    });
-  });
-
-  test('should maintain form input visual consistency', async ({ page }) => {
-    // Test form input elements for visual consistency
-    // Critical for user data entry experience
-    
-    const inputs = page.locator('input, textarea, select').first();
-    
-    if (await inputs.count() > 0) {
-      // Default state
-      await expect(inputs).toHaveScreenshot('world-creation-input-default.png', {
-        threshold: 0.2
-      });
-      
-      // Focused state
-      await inputs.focus();
-      await expect(inputs).toHaveScreenshot('world-creation-input-focus.png', {
-        threshold: 0.3
-      });
-      
-      // With content
-      await inputs.fill('Test World Name');
-      await expect(inputs).toHaveScreenshot('world-creation-input-filled.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain wizard navigation buttons visual consistency', async ({ page }) => {
-    // Test navigation buttons (Next, Previous, etc.)
-    // Essential for wizard progression
-    
-    const nextButton = page.locator('button:has-text("Next"), [data-testid="next-button"], button[type="submit"]').first();
-    const prevButton = page.locator('button:has-text("Previous"), button:has-text("Back"), [data-testid="previous-button"]').first();
-    
-    if (await nextButton.count() > 0) {
-      await expect(nextButton).toHaveScreenshot('world-creation-next-button.png', {
-        threshold: 0.2
-      });
-    }
-    
-    if (await prevButton.count() > 0) {
-      await expect(prevButton).toHaveScreenshot('world-creation-previous-button.png', {
-        threshold: 0.2
-      });
-    }
-  });
-
-  test('should maintain template selection visual layout', async ({ page }) => {
-    // Test template/preset selection interface if available
-    // Important for user experience in choosing world types
-    
-    const templateSelector = page.locator(
-      '[data-testid="template-selector"], .template-grid, .world-templates, .preset-selection'
-    ).first();
-    
-    if (await templateSelector.count() > 0) {
-      await expect(templateSelector).toHaveScreenshot('world-creation-template-selection.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for any grid or list of options
-      const optionGrid = page.locator('.grid, .options, .selection-grid').first();
-      if (await optionGrid.count() > 0) {
-        await expect(optionGrid).toHaveScreenshot('world-creation-options-grid.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-});
-
-test.describe('World Creation Form Validation Visual Tests', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/world/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Disable animations
-    await page.addStyleTag({
-      content: `
-        *, *::before, *::after {
-          animation-duration: 0s !important;
-          transition-duration: 0s !important;
-        }
-      `
-    });
-  });
-
-  test('should maintain validation error visual consistency', async ({ page }) => {
-    // Test visual appearance of validation errors
-    // Critical for user feedback during form submission
-    
-    // Try to trigger validation by submitting empty form
-    const submitButton = page.locator('button[type="submit"], button:has-text("Next"), button:has-text("Create")').first();
-    
-    if (await submitButton.count() > 0) {
-      await submitButton.click();
-      await page.waitForTimeout(100); // Allow validation to appear
-      
-      // Look for validation messages
-      const errorMessages = page.locator('.error, [role="alert"], .validation-error, .field-error');
-      
-      if (await errorMessages.count() > 0) {
-        await expect(errorMessages.first()).toHaveScreenshot('world-creation-validation-error.png', {
-          threshold: 0.25
-        });
-      }
-      
-      // Capture the form with validation errors
-      await expect(page).toHaveScreenshot('world-creation-form-with-errors.png', {
-        fullPage: true,
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain required field indicators visual consistency', async ({ page }) => {
-    // Test visual indicators for required fields
-    // Important for accessibility and user guidance
-    
-    const requiredFields = page.locator('input[required], textarea[required], select[required]');
-    const requiredIndicators = page.locator('.required, [aria-required="true"]');
-    
-    if (await requiredFields.count() > 0 || await requiredIndicators.count() > 0) {
-      const firstRequired = await requiredFields.count() > 0 
-        ? requiredFields.first() 
-        : requiredIndicators.first();
-        
-      await expect(firstRequired).toHaveScreenshot('world-creation-required-field.png', {
-        threshold: 0.2
-      });
-    }
-  });
-
-  test('should maintain success state visual consistency', async ({ page }) => {
-    // Test success/completion states if reachable
-    // Important for positive user feedback
-    
-    // Try to fill out a minimal valid form
-    const nameInput = page.locator('input[name*="name"], input[placeholder*="name"], #name').first();
-    
-    if (await nameInput.count() > 0) {
-      await nameInput.fill('Test World');
-      
-      // Look for success indicators or positive feedback
-      const successElements = page.locator('.success, .valid, [data-testid="success"]');
-      
-      if (await successElements.count() > 0) {
-        await expect(successElements.first()).toHaveScreenshot('world-creation-success-indicator.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-});
-
-test.describe('World Creation Image Generation Visual Tests', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto('/world/create');
-    await page.waitForLoadState('networkidle');
-  });
-
-  test('should maintain image generation interface visual consistency', async ({ page }) => {
-    // Test image generation UI if present
-    // Important for AI-powered world creation features
-    
-    const imageSection = page.locator(
-      '[data-testid="image-generation"], .image-generator, .world-image, .image-upload'
-    ).first();
-    
-    if (await imageSection.count() > 0) {
-      await expect(imageSection).toHaveScreenshot('world-creation-image-section.png', {
-        threshold: 0.3
-      });
-    } else {
-      // Look for any image-related elements
-      const imageElements = page.locator('img, .image-placeholder, [role="img"]');
-      
-      if (await imageElements.count() > 0) {
-        await expect(imageElements.first()).toHaveScreenshot('world-creation-image-element.png', {
-          threshold: 0.3
-        });
-      }
-    }
-  });
-
-  test('should maintain image upload visual states', async ({ page }) => {
-    // Test different states of image upload/generation
-    const uploadArea = page.locator('[data-testid="image-upload"], .upload-area, .dropzone').first();
-    
-    if (await uploadArea.count() > 0) {
-      // Default state
-      await expect(uploadArea).toHaveScreenshot('world-creation-upload-default.png', {
-        threshold: 0.25
-      });
-      
-      // Hover state
-      await uploadArea.hover();
-      await expect(uploadArea).toHaveScreenshot('world-creation-upload-hover.png', {
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain loading state during image generation', async ({ page }) => {
-    // Test loading states during AI image generation
-    const generateButton = page.locator('button:has-text("Generate"), [data-testid="generate-image"]').first();
-    
-    if (await generateButton.count() > 0) {
-      // Mock the image generation request to capture loading state
-      await page.route('**/api/generate-world-image', route => {
-        // Delay response to capture loading state
-        setTimeout(() => {
-          route.fulfill({
-            status: 200,
-            contentType: 'application/json',
-            body: JSON.stringify({ imageUrl: 'data:image/svg+xml;base64,PHN2Zz48L3N2Zz4=' })
-          });
-        }, 1000);
-      });
-      
-      await generateButton.click();
-      
-      // Capture loading state
-      const loadingElement = page.locator('.loading, .spinner, [data-testid="loading"]').first();
-      if (await loadingElement.isVisible()) {
-        await expect(loadingElement).toHaveScreenshot('world-creation-image-loading.png', {
-          threshold: 0.4
-        });
-      }
-    }
-  });
-});
-
-test.describe('World Creation Multi-Step Visual Flow', () => {
-  test('should maintain visual consistency across wizard steps', async ({ page }) => {
-    // Test the complete multi-step wizard flow
-    // Critical for overall user experience
-    
-    await page.goto('/world/create');
-    await page.waitForLoadState('networkidle');
-    
-    // Capture initial step
-    await expect(page).toHaveScreenshot('world-creation-step-1.png', {
-      fullPage: true,
-      threshold: 0.3
-    });
-    
-    // Try to progress through steps if possible
-    const nextButton = page.locator('button:has-text("Next"), [data-testid="next-button"]').first();
-    
-    if (await nextButton.count() > 0) {
-      // Fill minimal required data to progress
-      const nameInput = page.locator('input[name*="name"], input[placeholder*="name"]').first();
-      if (await nameInput.count() > 0) {
-        await nameInput.fill('Test World');
-      }
-      
-      await nextButton.click();
-      await page.waitForTimeout(200); // Allow step transition
-      
-      // Capture second step
-      await expect(page).toHaveScreenshot('world-creation-step-2.png', {
-        fullPage: true,
-        threshold: 0.3
-      });
-    }
-  });
-
-  test('should maintain wizard progress indicator throughout flow', async ({ page }) => {
-    // Test that progress indicators update correctly across steps
-    await page.goto('/world/create');
-    await page.waitForLoadState('networkidle');
-    
-    const progressElement = page.locator(
-      '[role="progressbar"], .progress, .step-indicator, .wizard-steps'
-    ).first();
-    
-    if (await progressElement.count() > 0) {
-      // Capture progress at different steps
-      await expect(progressElement).toHaveScreenshot('world-creation-progress-step-1.png', {
-        threshold: 0.25
-      });
-      
-      // Try to advance step and capture progress update
-      const nextButton = page.locator('button:has-text("Next")').first();
-      if (await nextButton.count() > 0) {
-        const nameInput = page.locator('input').first();
-        if (await nameInput.count() > 0) {
-          await nameInput.fill('Test');
-        }
-        
-        await nextButton.click();
-        await page.waitForTimeout(200);
-        
-        await expect(progressElement).toHaveScreenshot('world-creation-progress-step-2.png', {
-          threshold: 0.25
-        });
-      }
-    }
-  });
-});
